backend/pyproject.toml
[project]
name = "fairweaver"
version = "0.1.0"
description = "AI-assisted metadata interoperability platform with selectable pivot"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",
    "pyyaml==6.0.2",
    "httpx==0.27.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "httpx>=0.27",
    "ruff>=0.6",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "httpx>=0.27.2",
    "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
